finished implementation

+++ HW2/P3/driver.py
@@ -22,7 +22,8 @@ def make_coords(center=(-0.575 - 0.575j),
 
     x = np.linspace(start=(-width / 2), stop=(width / 2), num=count)
     xx = center + (x + 1j * x[:, np.newaxis]).astype(np.complex64)
-    return xx, np.zeros_like(xx, dtype=np.uint32)
+    #  use float 32 to more easily copy data from avx float8s
+    return xx, np.zeros_like(xx, dtype=np.float32)
 
 
 if __name__ == '__main__':
